+++ changelog.md
@@ -9,3 +9,8 @@
 * Fixed TCP Server delay while connecting (within ModbusLib subproject)
 * Added setting 'Max Connections' for max active TCP server connections
 * Update ModbusLib subproject up to v0.4.3
+
+# 0.2.0
+
+* Added unit list param (-sunit) for server to responde like '1,3,6-10,11,27'
+* Update ModbusLib subproject up to v0.4.4